+++ src/components/Input/Input.tsx
@@ -1,42 +1,37 @@
 import classNames from "classnames";
-import { useContext, type ComponentProps, type CSSProperties, type FC } from "react";
+import { merge } from "es-toolkit";
+import { type ComponentProps, type CSSProperties, type FC } from "react";
 
-import { ThemeContext } from "../ThemeProvider";
+import { useTheme } from "../ThemeProvider";
+import { InputBase } from "../InputBase";
 
 import styles from "./Input.module.scss";
-import { merge } from "es-toolkit";
 
 export type InputProps = Omit<ComponentProps<'input'>, 'size'> & {
   size?: 'sm' | 'md' | 'lg';
 };
 
 export const Input: FC<InputProps> = (props) => {
-  const theme = useContext(ThemeContext);
+  const theme = useTheme();
 
   const { size = 'md', className, style, ...rest } = props;
 
   const computedStyle = mergeStyles([
     size === 'sm' && {
-      '--input-height': theme.tokens.inputBaseSmHeight,
       '--input-font-size': theme.tokens.inputBaseSmFontSize,
       '--input-line-height': theme.tokens.inputBaseSmLineHeight,
     } as CSSProperties,
     size === 'md' && {
-      '--input-height': theme.tokens.inputBaseMdHeight,
       '--input-font-size': theme.tokens.inputBaseMdFontSize,
       '--input-line-height': theme.tokens.inputBaseMdLineHeight,
     } as CSSProperties,
     size === 'lg' && {
-      '--input-height': theme.tokens.inputBaseLgHeight,
       '--input-font-size': theme.tokens.inputBaseLgFontSize,
       '--input-line-height': theme.tokens.inputBaseLgLineHeight,
     } as CSSProperties,
     {
-      '--input-min-width': '16rem',
       '--input-padding-inline': '0.5rem',
-      '--input-border-color': theme.tokens.inputBaseDefaultBorderColor,
       '--input-border-color-focus': theme.colors.blue['800'],
-      '--input-border-radius': theme.tokens.inputBaseBorderRadius,
       '--input-placeholder-color': theme.tokens.inputBasePlaceholderColor,
       '--input-caret-color': theme.colors.blue['600'],
     } as CSSProperties,
@@ -44,11 +39,13 @@ export const Input: FC<InputProps> = (props) => {
   ]);
 
   return (
-    <input
-      className={classNames(className, styles.input)}
-      style={computedStyle}
-      {...rest}
-    />
+    <InputBase size={size}>
+      <input
+        className={classNames(className, styles.input)}
+        style={computedStyle}
+        {...rest}
+      />
+    </InputBase>
   );
 }