+++ src/components/Input/CHANGELOG.md
@@ -5,3 +5,4 @@
 - 2025-09-22: Support the `size` props (`sm`, `md`, `lg`) for Input component to
   adjust height and font size. Etc.
 - 2025-09-26: Use theme tokens for Input component styles.
+- 2025-09-30: Use `InputBase` as a wrapper for the Input component.